Update Library title to eungabi, add copyright header annotation
