--- androidApp/src/main/java/com/blucky8649/decompose_navhost/android/MainActivity.kt
@@ -1,42 +1,0 @@
-package com.blucky8649.decompose_navhost.android
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.*
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
-import com.arkivanov.decompose.ExperimentalDecomposeApi
-import com.arkivanov.decompose.defaultComponentContext
-import com.blucky8649.decompose_navhost.SampleApp
-import com.blucky8649.decompose_navhost.navigation.NavHost
-import com.blucky8649.decompose_navhost.navigation.rememberNavController
-
-class MainActivity : ComponentActivity() {
-    @OptIn(ExperimentalDecomposeApi::class)
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContent {
-            DecomposeSampleTheme {
-                SampleApp(componentContext = defaultComponentContext())
-            }
-        }
-    }
-}
-
-@Composable
-fun GreetingView(text: String) {
-    Text(text = text)
-}
-
-@Preview
-@Composable
-fun DefaultPreview() {
-    DecomposeSampleTheme {
-        GreetingView("Hello, Android!")
-    }
-}
