+++ androidApp/src/main/java/com/blucky8649/eungabi/android/MainActivity.kt
@@ -1,0 +1,31 @@
+package com.blucky8649.eungabi.android
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.compose.material3.*
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.tooling.preview.Preview
+import com.blucky8649.eungabi.SampleApp
+
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContent {
+            SampleApp()
+        }
+    }
+}
+
+@Composable
+fun GreetingView(text: String) {
+    Text(text = text)
+}
+
+@Preview
+@Composable
+fun DefaultPreview() {
+    DecomposeSampleTheme {
+        GreetingView("Hello, Android!")
+    }
+}
