--- decompose-navhost/src/androidMain/kotlin/com/blucky8649/decompose_navhost/utils/Uuid.android.kt
@@ -1,5 +1,0 @@
-package com.blucky8649.decompose_navhost.utils
-
-import java.util.UUID
-
-actual val randomUuid: String get() = UUID.randomUUID().toString()
