--- decompose-navhost/src/commonMain/kotlin/com/blucky8649/decompose_navhost/navigation/PopUpTo.kt
@@ -1,5 +1,0 @@
-package com.blucky8649.decompose_navhost.navigation
-
-class PopUpToBuilder {
-    var inclusive: Boolean = false
-}
