--- decompose-navhost/src/commonMain/kotlin/com/blucky8649/decompose_navhost/utils/NavUtils.kt
@@ -1,5 +1,0 @@
-package com.blucky8649.decompose_navhost.utils
-
-import com.blucky8649.decompose_navhost.navigation.SCHEME
-
-fun withScheme(route: String) = if (route.startsWith(SCHEME)) route else "$SCHEME$route"
