--- decompose-navhost/src/commonMain/kotlin/com/blucky8649/decompose_navhost/utils/Uuid.kt
@@ -1,3 +1,0 @@
-package com.blucky8649.decompose_navhost.utils
-
-expect val randomUuid: String
