--- decompose-navhost/src/iosMain/kotlin/com/blucky8649/decompose_navhost/utils/Uuid.ios.kt
@@ -1,5 +1,0 @@
-package com.blucky8649.decompose_navhost.utils
-
-import platform.Foundation.NSUUID
-
-actual val randomUuid: String get() = NSUUID.UUID().UUIDString
