+++ eungabi/src/commonMain/kotlin/com/blucky8649/eungabi/navigation/experimental/EunGabiState.kt
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2024 easternkite
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.blucky8649.eungabi.navigation.experimental
+
+import com.blucky8649.eungabi.navigation.NavBackStackEntry
+
+internal object EunGabiState {
+    private var backQueue = ArrayDeque<NavBackStackEntry>()
+
+    fun save(backQueue: ArrayDeque<NavBackStackEntry>) {
+        this.backQueue.addAll(backQueue)
+    }
+
+    fun restore(): ArrayDeque<NavBackStackEntry> {
+        val result = ArrayDeque(backQueue.toList())
+        return result.also { backQueue.clear() }
+    }
+}
