+++ eungabi/src/commonMain/kotlin/com/blucky8649/eungabi/utils/NavUtils.kt
@@ -1,0 +1,20 @@
+/*
+ * Copyright 2024 easternkite
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.blucky8649.eungabi.utils
+
+import com.blucky8649.eungabi.navigation.SCHEME
+
+fun withScheme(route: String) = if (route.startsWith(SCHEME)) route else "$SCHEME$route"
