+++ gradle/libs.versions.toml
@@ -5,8 +5,6 @@ coreKtx = "1.13.1"
 compose = "1.7.2"
 compose-material3 = "1.3.0"
 cmp-plugin = "1.7.0-beta02"
-decompose = "3.2.0-beta01"
-essenty = "2.2.0-beta01"
 androidx-activityCompose = "1.9.2"
 lifecycleRuntimeKtx = "2.8.6"
 cupertino = "0.1.0-alpha04"
@@ -25,10 +23,6 @@ compose-ui-tooling-preview = { module = "androidx.compose.ui:ui-tooling-preview"
 compose-foundation = { module = "androidx.compose.foundation:foundation", version.ref = "compose" }
 compose-material3 = { module = "androidx.compose.material3:material3", version.ref = "compose-material3" }
 
-decompose = { module = "com.arkivanov.decompose:decompose", version.ref = "decompose" }
-decompose-extension-compose = { module = "com.arkivanov.decompose:extensions-compose", version.ref = "decompose" }
-decompose-extension-compose-experimental = { module = "com.arkivanov.decompose:extensions-compose-experimental", version.ref = "decompose" }
-decompose-extension-android = { module = "com.arkivanov.decompose:extensions-android", version.ref = "decompose" }
 uri = { module = "com.eygraber:uri-kmp", version.ref = "uri-kmp" }
 napier = { module = "io.github.aakira:napier", version.ref = "napier"}
