--- shared/src/commonMain/kotlin/com/blucky8649/decompose_navhost/DefaultPlatformAnimation.kt
@@ -1,11 +1,0 @@
-package com.blucky8649.decompose_navhost
-
-import com.arkivanov.decompose.ExperimentalDecomposeApi
-import com.arkivanov.decompose.extensions.compose.experimental.stack.animation.StackAnimation
-import com.arkivanov.essenty.backhandler.BackHandler
-
-@OptIn(ExperimentalDecomposeApi::class)
-internal expect fun defaultPlatformAnimation(
-    backHandler: BackHandler,
-    onBack: () -> Unit
-) : StackAnimation<Any, Any>
