+++ shared/src/iosMain/kotlin/com/blucky8649/eungabi/ViewController.kt
@@ -1,0 +1,7 @@
+package com.blucky8649.eungabi
+
+import androidx.compose.ui.window.ComposeUIViewController
+
+fun SampleViewController() = ComposeUIViewController {
+    SampleApp()
+}
